Refactoring WriteTestData method and adding PayloadReader assertions
